misc. warnings

+++ TheRegularEngineV2/TheRegularEngine/Graphics/FontRenderer.cpp
@@ -95,7 +95,7 @@ namespace TRE
 			if (TextComp.m_IsDialogue) //If dialogue we handle what to print here
 			{
 				TextComp.m_Timer += Engine::GetInstance().GetWindow()->GetDeltaTime();
-				int RenderSize = TextComp.m_Timer * TextComp.m_Speed;
+				int RenderSize = static_cast<int>(TextComp.m_Timer * TextComp.m_Speed);
 				if (RenderSize < TextComp.m_TextContent.Text.size())
 					TextToRender = TextComp.m_TextContent.Text.substr(0, RenderSize);
 				else
